Fix docker-compose port mapping
Prior to this commit, port maps in the docker-compose file weren't
sending traffic to the internal ports that the applications
are expecting to recieve traffic on.

This commit maps ports correctly so that a local environment doesn't
have to match port numbers that the containers are expecting the find
traffic on.

Closes #29