Make `first_name`, `last_name`, and `term_end` not null
Prior to this commit, we have 5 or so rows in our `member_of_congress`
table that didn't have this information populated. This is because
this data is stale.

We now have an automated system to keep our MoC data updated. Rows where
this data was null have resigned/died/left congress prior to this
automation being in place.

We do not currently have any data referencing these members of congress,
so deleting their data won't cascade.