Only select currently serving Members of Congress
Now that we're tracking more Members of Congress than are currently
serving, we need to use `member_of_congress.term_end` to determind if a
given member is currently serving.

We should only be selecting MoC who are currently serving, as they're
the people who are meant to recieve correspondence.